Update patch set 14
Patch Set 14:

I understand that there is some cost to posting the message to a thread's ring, but there are a few things to keep in mind:

1) If the populate or commit flags are set to true, a bdev module may need to perform an asynchronous operation. For instance, say a bdev module does DRAM caching and when initiating a zcopy, there is a cache miss. I wanted the design to handle this case.

2) The atomic instruction to insert into the ring does have a cost, but cross-thread message passing is used very sparingly throughout SPDK so the actual memory location that the atomic is acting on is not heavily contended and is almost certainly already in the local CPU's cache. I do understand the different between lock-free and wait-free and I'm sensitive to the impact that has on P99+ latency, but I'm struggling to believe that this shows up as a problem in a real benchmark. As with anything, I can be convinced if there is data.

3) Even if real data can be provided that shows it is a problem, the implementation of the underlying message passing code could easily optimize for this case by having a local SP/SC ring used for passing messages only to itself.

Is the asynchronous nature of the new zcopy API, given that it performs a single atomic instruction on a ring, the primary complaint? I've been really trying to divine the nature of the push back and haven't gotten any technical reasons until this one.

Patch-set: 14
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>