Update patch set 14
Patch Set 14:

Reusing the bdev descriptor is an interesting idea.  But I'm not sure it works - there's context we need at the scsi layer to know when the last descriptor has been closed.

Maybe we add an API to the bdev to get a count of number of open descriptors?  I don't particularly care for it, but maybe it's worth exploring.

But one problem is that it then completely precludes a bdev from being shared by an iSCSI target node and another upper layer construct (vhost-blk, NVMe-oF).

I'm leaning towards sticking with the new scsi_lun descriptor for now.

Patch-set: 14
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>