Update patch set 14
Patch Set 14: Code-Review+2

The other reason that open/close is really needed here is so we register the hot remove callback from a different place.  Currently it's done when the lun is constructed, which totally does not work for iSCSI.  When the lun gets hot removed, you have to go search all of the open sessions to find out which one might be using the removed LUN.  It's a mess (and I know - I tried implementing it a few months ago).

How is an r2t task a SCSI task though?  I really don't see that.  Yes, we store information about the original SCSI request as an spdk_scsi_task, but that task never gets submitted to the SCSI layer.  It's purely to track where we are as the client starts responding to our R2T requests with Data Out PDUs.  It's the Data Out PDUs that get converted to scsi_tasks which are actually submitted to the SCSI layer.

Patch-set: 14
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2