Update patch set 14
Patch Set 14:

Answering this question:
Is the asynchronous nature of the new zcopy API, given that it performs a single atomic instruction on a ring, the primary complaint? I've been really trying to divine the nature of the push back and haven't gotten any technical reasons until this one.

Sriram pointed out the issue with the context switch. I'll add a few more points about this change.

1. The first issue is with the ZCOPY semantics. ZCOPY is not a type of BDEV IO. Read/write/flush/write-zeros are all types of IO's. ZCOPY is an operation. Eg. You can achieve a Read IO with or without ZCOPY. So, I would consider a zero-copy semantic as a "method/operation" that can be performed on a IO class. Different IO types may or may not support ZCOPY. Same case with different IO's on different BDEVs. This is my first technical issue with this change.

2. At least initially, the semantics of Zcopy was not intuitive (to me). Without the examples, it was hard to see how the populate and the commit flags are being used. This issue may not be considered technical in nature. 

3. The second technical issue is the context switch. I want to add a few more points to the questions Ben asked.
a. Cross-thread message passing is used very sparingly in SPDK. Yes, but that is not issue here. This queue gets used for every IO when there is an asynchronous backend. Every backend callback (for every IO) gets deposited in this queue. So, it is used heavily and will be contended heavily as well.
b. Let us also consider the context switch issue with the virtual thread functionality coming in. This change introduces latency as follows:
i.  A virtual thread can have multiple pollers in its ring.
ii. The poll function is called for this virtual thread. It first looks at the message/event queue. Then it starts working on the polling ring. It picks the first poller out. This results in an addition to the message queue for the zcopy cb. But, the thread does not immediately process it. It moves on to the next IO in that queue. It then moves to the next poller function in that ring. It then handles the timers. The zcopy cb is still stuck in the message queue.
iii. The physical thread may then pick the next virtual thread to work upon after one iteration of the poll function. That means, this zcopy callback has to now wait for this virtual thread to be "rescheduled" back on some physical thread before it can be processed.

Patch-set: 14
Reviewer: Madhu Pai <1014342@d5d70762-12d0-45a1-890d-524b12d3f735>