Update patch set 14
Patch Set 14:

> 1. The first issue is with the ZCOPY semantics. ZCOPY is not a type of BDEV IO.
> Read/write/flush/write-zeros are all types of IO's. ZCOPY is an operation. Eg. You can achieve
> a Read IO with or without ZCOPY. So, I would consider a zero-copy semantic as a
> "method/operation" that can be performed on a IO class. Different IO types may or may not
> support ZCOPY. Same case with different IO's on different BDEVs. This is my first technical 
> issue with this change.

ZCOPY is inherently different though - especially for writes.  So the API needs to change to support it.  The API as Ben has it is much more flexible  - it allows both READ and WRITE for the same IO!  This may not be needed for NVMe-oF, but there are other use cases for it (i.e. mmap-like operations for filesystems).

I'm still not sure I understand the technical issue though.  It's trivial to determine if the ZCOPY IO is a READ for your case - just check the populate flag.  Same for WRITE - just check the commit flag.

> 3. The second technical issue is the context switch. I want to add a few more points to the
> questions Ben asked.
> a. Cross-thread message passing is used very sparingly in SPDK. Yes, but that is not issue
> here. This queue gets used for every IO when there is an asynchronous backend. Every
> backend callback (for every IO) gets deposited in this queue. So, it is used heavily and will be
> contended heavily as well.

But what's the measured performance impact?  If you have data and code that you can share that indicates this contention matters, please share.  I don't see where we would have contention.  We can easily modify SPDK to have a special queue for messages that a thread passes to itself.

> b. Let us also consider the context switch issue with the virtual thread functionality coming in.
> This change introduces latency as follows:
> i.  A virtual thread can have multiple pollers in its ring.
> ii. The poll function is called for this virtual thread. It first looks at the message/event queue.
> Then it starts working on the polling ring. It picks the first poller out. This results in an addition
> to the message queue for the zcopy cb. But, the thread does not immediately process it. It
> moves on to the next IO in that queue. It then moves to the next poller function in that ring. It > then handles the timers. The zcopy cb is still stuck in the message queue.
> iii. The physical thread may then pick the next virtual thread to work upon after one iteration
> of the poll function. That means, this zcopy callback has to now wait for this virtual thread to
> be "rescheduled" back on some physical thread before it can be processed.

This is a valid point.  We will need to make sure that after all pollers for a thread are executed, that we execute any pending messages from that thread before it cedes control for another virtual thread.

Patch-set: 14
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>