Update patch set 14
Patch Set 14:

> Patch Set 14:
> 
> This just a note to say that NetApp would like all work on this patch series to stop until there has been further communication about what the NetApp and Broadcom requirements are w.r.t. zero copy and DMA support.

John - keep in mind that often times in communities like this its easier to explain things in code via patch reviews so it probably doesn't make sense to 'stop', this is part of the process in fleshing something out that works for everyone. We should continue to iterate here, on the mailing list, on IRC and in community meetings.  Just make sure with every patch set you put your -1 on it indicating that you don't believe it's ready yet.  That will be respected, fear not :)

Patch-set: 14
Reviewer: Paul Luse <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>