Update patch set 14
Patch Set 14:

Hi Madhu,

The 'special' path for a thread to pass messages to itself would be generally helpful to SPDK.  This conversation is expediting the need for it, but I don't consider it special case.  We can make this part of the process super optimized - per-thread caches of event objects and only using the backing mempool in this case when the cache is exhausted.

Regarding running posting events one last time before ceding control back to the scheduler/reactor - this is something we probably would have done anyways as we move to this threading model.  This was a great point and something I hadn't thought of yet.

I still think the async version of these zcopy routines are mandatory - even for writes.  I guess NetApp doesn't have a case where a write operation has to wait for a DRAM buffer (i.e. cache is full and something has to be flushed to disk) but I don't see that as universal.  So I think the question is do we need synchronous versions of the zcopy routines - and if so, what would the API look like and how would bdev modules be modified to support it.

-Jim

Patch-set: 14
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>