Update patch set 14
Patch Set 14: -Code-Review

> Hi Madhu,
 > 
 > The 'special' path for a thread to pass messages to itself would be
 > generally helpful to SPDK.  This conversation is expediting the
 > need for it, but I don't consider it special case.  We can make
 > this part of the process super optimized - per-thread caches of
 > event objects and only using the backing mempool in this case when
 > the cache is exhausted.
 > 
 > Regarding running posting events one last time before ceding
 > control back to the scheduler/reactor - this is something we
 > probably would have done anyways as we move to this threading
 > model.  This was a great point and something I hadn't thought of
 > yet.

Fair points. FWIW, the reason I have accessor functions for the virtual thread in my patch is for this very reason. I wanted the app to control the way polling, timer and message handling is done and not the library. The reason was that we already do this - we look at the message queue once more before we swap out the virtual thread from the physical thread.

 > 
 > I still think the async version of these zcopy routines are
 > mandatory - even for writes.  I guess NetApp doesn't have a case
 > where a write operation has to wait for a DRAM buffer (i.e. cache
 > is full and something has to be flushed to disk) but I don't see
 > that as universal.  So I think the question is do we need
 > synchronous versions of the zcopy routines - and if so, what would
 > the API look like and how would bdev modules be modified to support
 > it.
 > 
 > -Jim

Sriram answered this just now. We do have situations where we dont have the DRAM buffers available for writes. But, in such cases we do not want to allocate the bdev_io. We return a status that tells the transport that enough resources are NOT available to complete the processing of the IO. Transport would then push this into the pending queue.
Basically, the init function of an IO type decides if all needed resources for that IO are available. If not, we want to park that IO.

Patch-set: 14
Reviewer: Madhu Pai <1014342@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0