Update patch set 14
Patch Set 14:

> Patch Set 14:
> 
> Certainly, stopping would be a huge mistake. We're finally getting down to the technical discussion that's been so badly needed. As the others have said, we won't merge until everyone's needs are met. This is how collaborative development works.
> 
> (Also I'm very interested to hear an answer to Jim's question - is there really no scenario on the write path where the zcopy start may not have any memory available? That seems fishy to me.)

We also need to design when the bdev module needs to give up if buffers are not available. It can't wait forever too.
I was wondering how do you handle if the memory is available at some time later and make sure aborts are handled properly.
You can't stitch those iovs to bdev_io in the a different thread so you need to post an event with those buffer/iovs/context to the poller thread. Meanwhile if we get an abort? How do you determine the IO is waiting in bdev or the bdev module gave up because buffers are not available.
From abort context you have a req and bdev_io. You don't know if its held up in bdev module waiting for buffers or bdev module failed it(this is not handled today) or if its race where the buffers are in queue while the abort is handled.

Patch-set: 14
Reviewer: Sriram Popuri <1011327@d5d70762-12d0-45a1-890d-524b12d3f735>