Update patch set 14
Patch Set 14:

>> Madhu: We do have situations where we dont have the DRAM buffers available for writes. But, in such cases we do not want to allocate the bdev_io. We return a status that tells the transport that enough resources are NOT available to complete the processing of the IO. Transport would then push this into the pending queue.

This isn't the optimal behavior when the bdev is out of resources because it gives the upper layer no mechanism to proceed other than periodically retrying. What you actually want to do is trigger a notification when the resource becomes available and use that to resume processing the new request. In other words, if the bdev module is out of resources, it should do the queueing, not the transport. That's why, in these error conditions, an asynchronous zcopy API is a superior design.

To summarize my understanding here:
1) For both read and write, the worst case *does* require the zcopy start and end paths to be asynchronous.
2) For write (populate == false), the vast majority of the time the zcopy start will complete synchronously.
3) For read (commit == false), the vast majority of the time the zcopy end will complete synchronously.
4) NetApp believes that there is a big enough performance benefit to make it worthwhile to add special handling to optimize for the case where the call could be done synchronously.

I originally coded it the way I did to keep it simple while still handling all possible scenarios, not necessarily thinking point #4 was worth making the API significantly more complex. But let's say I'm convinced and begin building from there.

We could certainly take some inspiration from Srikanth's design (although that was specific to read) and make the API look as follows:

int spdk_bdev_zcopy_start(struct spdk_bdev_desc *desc,
                          struct spdk_io_channel *ch,
			  uint64_t offset_blocks, uint64_t num_blocks,
			  bool populate,
                          struct spdk_bdev_io **bdev_io, // output
			  spdk_bdev_io_completion_cb cb, void *cb_arg);

When spdk_bdev_zcopy_start is called, if successful it fills in the bdev_io parameter that the user provided. The user can then check the status on the bdev_io via a new function that will return the already existing bdev_io status, which is either pending (if the call will execute asynchronously) or success/failure (if done synchronously). In the case where the call completes synchronously, the provided callback will not be called. I've confirmed that there is a fairly simple way in the code to make this happen.

Similarly for zcopy end:

int spdk_bdev_zcopy_end(struct spdk_bdev_io *bdev_io,
                        bool commit,
			spdk_bdev_io_completion_cb cb, void *cb_arg);

And if it succeeds, the user can check the status of the bdev_io to see if it is pending or already done.

Thoughts?

Patch-set: 14
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>