Update patch set 14
Patch Set 14:

> Patch Set 14:
> 
> >> Madhu: We do have situations where we dont have the DRAM buffers available for writes. But, in such cases we do not want to allocate the bdev_io. We return a status that tells the transport that enough resources are NOT available to complete the processing of the IO. Transport would then push this into the pending queue.
> 
> This isn't the optimal behavior when the bdev is out of resources because it gives the upper layer no mechanism to proceed other than periodically retrying. What you actually want to do is trigger a notification when the resource becomes available and use that to resume processing the new request. In other words, if the bdev module is out of resources, it should do the queueing, not the transport. That's why, in these error conditions, an asynchronous zcopy API is a superior design.

I don't agree with this. I feel queueing works at protocol layer and not bdev module because the req timeouts are defined there and host can escalate to controller rest or LOGO if its not honored. So the protocol(either in transport or target) layer is the best to decide how long to retry an IO internally. I agree if there is notification from bdev modue on when to retry can help, but for a simple solution its easy to retry at a periodic intervals before giving up after a fixed time budget which should is within the protocol timeouts.

> 
> To summarize my understanding here:
> 1) For both read and write, the worst case *does* require the zcopy start and end paths to be asynchronous.
> 2) For write (populate == false), the vast majority of the time the zcopy start will complete synchronously.
> 3) For read (commit == false), the vast majority of the time the zcopy end will complete synchronously.
> 4) NetApp believes that there is a big enough performance benefit to make it worthwhile to add special handling to optimize for the case where the call could be done synchronously.
> 
Your understanding is correct w.r.t. how NetApp handles zcopy except point 3 which is always synchronous. For 2 also we handle it synchronously and if DRAM buffers are full its a failure. How the failure is handled is upto the protocol layer.

> I originally coded it the way I did to keep it simple while still handling all possible scenarios, not necessarily thinking point #4 was worth making the API significantly more complex. But let's say I'm convinced and begin building from there.
> 
> We could certainly take some inspiration from Srikanth's design (although that was specific to read) and make the API look as follows:
> 
> int spdk_bdev_zcopy_start(struct spdk_bdev_desc *desc,
>                           struct spdk_io_channel *ch,
> 			  uint64_t offset_blocks, uint64_t num_blocks,
> 			  bool populate,
>                           struct spdk_bdev_io **bdev_io, // output
> 			  spdk_bdev_io_completion_cb cb, void *cb_arg);
> 
> When spdk_bdev_zcopy_start is called, if successful it fills in the bdev_io parameter that the user provided. The user can then check the status on the bdev_io via a new function that will return the already existing bdev_io status, which is either pending (if the call will execute asynchronously) or success/failure (if done synchronously). In the case where the call completes synchronously, the provided callback will not be called. I've confirmed that there is a fairly simple way in the code to make this happen.
> 

So your proposal is the bdev_io is allocated/acquired in bdev module correct?
If so after the call to zcopy_start, the target can check if bdev_io is not NULL and proceed for R2T. If NULL it is considered asynchronous.

> Similarly for zcopy end:
> 
> int spdk_bdev_zcopy_end(struct spdk_bdev_io *bdev_io,
>                         bool commit,
> 			spdk_bdev_io_completion_cb cb, void *cb_arg);
> 
> And if it succeeds, the user can check the status of the bdev_io to see if it is pending or already done.
> 
> Thoughts?

Here it makes sense for a status function.

I get what you are saying. The bdev_io is acquired in bdev module, but to figure out if its pending or done, you use a new status check function always(where its write path or read path. Whether its during zcopy_start/end).

We want to acquire bdev_io and iovs in one context. Either we get both or none. The failure should be handled at protocol layer.

Patch-set: 14
Reviewer: Sriram Popuri <1011327@d5d70762-12d0-45a1-890d-524b12d3f735>