Update patch set 14
Patch Set 14:

I agree with both of you (Sriram and John). To recap, the points are:

1) The total amount of time to wait for an I/O to complete should be controlled by the protocol layer (in this case, NVMe-oF).
2) If the protocol layer is out of resources, it should be in charge of waiting for them to become available (i.e. FC should wait for its own data buffers or other resources).

However, neither of those things are what we're discussing here. We're discussing what to do when the bdev module runs out of resources (i.e. the "backend", the storage side, or whatever you want to call it), which is what this API is accessing. In that case, I'm saying it is a better design to allow the bdev module to do the wait intelligently, bounded by a total timeout enforced at the protocol layer, than to have the protocol layer periodically retry. This is why it is important that the zcopy path can be asynchronous.

Patch-set: 14
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>