Update patch set 14
Patch Set 14:

> Patch Set 14:
> 
> I agree with both of you (Sriram and John). To recap, the points are:
> 
> 1) The total amount of time to wait for an I/O to complete should be controlled by the protocol layer (in this case, NVMe-oF).
> 2) If the protocol layer is out of resources, it should be in charge of waiting for them to become available (i.e. FC should wait for its own data buffers or other resources).
> 
> However, neither of those things are what we're discussing here. We're discussing what to do when the bdev module runs out of resources (i.e. the "backend", the storage side, or whatever you want to call it), which is what this API is accessing. In that case, I'm saying it is a better design to allow the bdev module to do the wait intelligently, bounded by a total timeout enforced at the protocol layer, than to have the protocol layer periodically retry. This is why it is important that the zcopy path can be asynchronous.

Ben,
I have to disagree with you on this. You make it sound like the transport layer will retry on some sort of time periodic basis, which is not the case.  The transport layer will retry when and only when it gets a command completion callback. So it knows exactly when to retry beacuse that's the only time it will be possible when you are out of resources to try to get resources for a pending io (also after timeout or abort processing, which I won't get into here). Since buffers aren't the only resource FC needs, it makes more sense for the transport to handle the all cases of unavailable resources. So the zcopy in bdev layer should be very simple and only return "here's your buffers" or "I don't got enough buffers for you right now". Then the transport layer can put the request on a pending queue, so it can be pulled off and retried on a command complete callback later. This is how we are doing it today. Also, in the future when we add QoS, that pending queue will probably have to be prioritized, which will be more complicated to do in the bdev layer.

Patch-set: 14
Reviewer: John Barnard <1014365@d5d70762-12d0-45a1-890d-524b12d3f735>