Update patch set 14
Patch Set 14:

I've read your responses a number of times and with Seth's help I think I understand what one of the misunderstandings is.

John >> You make it sound like the transport layer will retry on some sort of time periodic basis, which is not the case.  The transport layer will retry when and only when it gets a command completion callback. So it knows exactly when to retry beacuse that's the only time it will be possible when you are out of resources to try to get resources for a pending io...

This statement is not correct. There are bdev module resources that may not be tied to the protocol's bdev_io and become available independently of I/O completions. For instance, imagine a caching bdev module that must evict some data to make room for a new write. The total queue depth at the time of the eviction could be 1 (i.e. just the incoming I/O), but the new write buffer doesn't become available until some internal operation completes. Therefore, if the bdev layer returns out of resources, the transport layer would have no recourse but to periodically retry. That's why the API must be asynchronous and allow the particular bdev module to hold the I/O until it completes. The transport can still place a global timeout on the I/O and send an abort if the whole operation is taking too long, but the details of waiting on resources are more complex than retrying on an I/O completion.

Patch-set: 14
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>