bdev/nvme : A valid construct_nvme_bdev request is failing after an invalid attempt
When the following steps are followed:
  1) Run construct_nvme_bdev with SSD1 and name nvme0 - This operation succeeds
  2) Run construct_nvme_bdev with SSD2 and name nvme0 - This result in error
  3) Run construct_nvme_bdev with SSD2 and name nvme1 - This results in error

In the step 3, the operation should ideally succeed, but it fails.

Fix:
The issue happens because in step 2, the SSD2 is already registered/probed and is already added to nvmf_tgt.
Hence in step 3, when it is added again, the probe/registration fails.

In this patch, we have added check before probing to ensure that the base name is not used earlier.
With this fix, in step 2, SSD2 will not be probed/registered. Hence in step 3, it will succeed.

Change-Id: Id025dc3c311646c872389ab48b299fd7084fa8a5
Signed-off-by: Rakesh Babu <Rakesh.Babu@wdc.com>