Update patch set 48
Patch Set 48:

Pawel, was just starting to rework some of the op creation code to deal with iovs that are < a block in length because we decided sorta late in the game that the IV for ciphers was going to be LBA so that's our smallest unit of crypt/decrypt.  Given that, I realize now that I can't do decryption in place anyway because the target buffer is not assured to be made of of unit of the block size of the underlying device (likely not). So I need to do what I'm doing with writes and decrypt into my own buffer an LBA size at a time then copy the data (or whatever zcopy cool shit we may have one of these days) to the a bdev allocated contiguous buffer. Hope that makes sense.

Patch-set: 48
Reviewer: Paul Luse <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>