Update patch set 1
Patch Set 1:

I have a question on this "The issue happens because in step 2, the SSD2 is already registered/probed and is already added to nvmf_tgt."

I haven't looked at the relevant code path but it seems that if step 2 results in error that SSD2 should be 'removed from the tgt' if that's a thing.

Overall it seems like less pre-checking is better (for performance, to keep it clean) and more effort should be placed on correctly handling and cleaning up after an error.

I could be totally off base here in this case though, please let me know!

Patch-set: 1
CC: Paul Luse <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>