nvmf/rdma: drain queue before deletion.
Previously, we simply deleted the queue pair without clearing all of the
pending requests. This could lead to undefined behavior in the deletion
path, such as receiving completions on the completion queue for a wr on
a deleted queue. By tracing all of the completions through the
different states, we can avoid accessing a qpair after it has been
deleted.

Change-Id: I4935627944bb8c10a1dc7ff0ac1301260491c61b
Signed-off-by: Seth Howell <seth.howell@intel.com>