bdev/raid: Use free() instead of realloc() to free allocated memory
Currently realloc() is used to free memory unintentionally. Hence
the pointer is not nullified and it has caused double free as reported
in Github issue #372.

Fix the issue in this patch, and use free() instead of realloc() when
the size is 0 because the behavior of realloc() when the size is 0 is
implementation specific.

Change-Id: I310e25965a5ac14a5c5b6aaa4f86514652332e80
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>