autobuild.sh: don't check for untracked files in submodules
This is a somewhat temporary workaround for avoiding git status
errors related to symlinked files in the DPDK submodule.  The
build pools will be fixed to not change symlinks to regular
files when copying repositories between systems, but until
then apply this fix so we can move forward with switching
to DPDK 18.05 (which now has symlinked files in its
repository).

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I8a6df76fb77a6f3ee3817bd975d198ae4cc3014a