Update patch set 5
Patch Set 5:

In RAID terminology, "strip" relates to single disk in a RAID group (array of disks) and is predefined number of contiguous addressable blocks in a single disk. "stripe" is mostly number of "strips" in a row. For e.g. in Raid0 raid group of 5 disks, it can be configured to "strip" size of 128KB, "stripe" size is number of disks multiply by strip size (5 * 128KB)

Patch-set: 5