bdev/iscsi: fix memory leaks on connection failures
create_iscsi_lun freed the provided parameters under
one condition, and did not under some other. The
behavior is now clarified.

When create_iscsi_lun succeeded, the parameters are
passed to an iscsi bdev and will be freed together
with its removal, but if the bdev could not be created,
the params need to be freed manually.

Change-Id: I010421d48461b365a324488029e11c90a4e20a37
Signed-off-by: Dariusz Stojaczyk <dariuszx.stojaczyk@intel.com>