Update patch set 1
Patch Set 1: Code-Review+1

This looks good to me, but one question I have is why do we keep and use the data in the bs object if we are going t override it in the superblock? Wouldn't it be useful to just keep one copy of each data point in either of these data structures?
Or is it for backwards compatibility in case the super block doesn't specify some of that information?

Patch-set: 1
Reviewer: Seth Howell <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1